// src/app/models/user.ts
export interface User {
  fname: string;
  lname: string;
  phone: string;
  email: string;
  date: string;
  time: string;
  numberOfPeople: number;
  numberOfTables: number;
  message: string;

  // id?: string;
  // foodId: string;
  // itemName: string;
  // price: number;
  // rating: number;
  // description: string;


}